extract duplication
