+++ apollo-openapi/pom.xml
@@ -38,6 +38,10 @@
       <artifactId>apollo-core</artifactId>
     </dependency>
     <dependency>
+      <groupId>com.ctrip.framework.apollo</groupId>
+      <artifactId>apollo-common</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.apache.httpcomponents</groupId>
       <artifactId>httpclient</artifactId>
       <exclusions>
